AE_CEL-378 - Improved structure of app "framework"
