+++ applications/common/appInit.h
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2022 u-blox
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+int32_t getSerialNumber(void);
+
+int32_t initCellularDevice(void);
+bool initXplrDevice(void);
+
+void dwellAppLoop(void);
+int32_t setAppDwellTime(commandParamsList_t *params);
+int32_t setAppLogLevel(commandParamsList_t *params);
+
+bool startupFramework(void);
+void finalise(applicationStates_t appState);
